requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
allure-pytest==2.13.5
    # via -r requirements.in
allure-python-commons==2.13.5
    # via allure-pytest
attrs==24.2.0
    # via
    #   allure-python-commons
    #   hypothesis
    #   jsonschema
    #   outcome
    #   referencing
    #   trio
certifi==2024.8.30
    # via
    #   requests
    #   selenium
cffi==1.17.0
    # via trio
charset-normalizer==3.3.2
    # via requests
colorama==0.4.6
    # via
    #   loguru
    #   pytest
et-xmlfile==1.1.0
    # via openpyxl
execnet==2.1.1
    # via pytest-xdist
faker==28.1.0
    # via -r requirements.in
h11==0.14.0
    # via wsproto
hypothesis==6.111.2
    # via -r requirements.in
idna==3.8
    # via
    #   requests
    #   trio
iniconfig==2.0.0
    # via pytest
jsonschema==4.23.0
    # via -r requirements.in
jsonschema-specifications==2023.12.1
    # via jsonschema
loguru==0.7.2
    # via -r requirements.in
openpyxl==3.1.5
    # via -r requirements.in
outcome==1.3.0.post0
    # via trio
packaging==24.1
    # via
    #   pipdeptree
    #   pytest
pipdeptree==2.23.1
    # via -r requirements.in
pluggy==1.5.0
    # via
    #   allure-python-commons
    #   pytest
pycparser==2.22
    # via cffi
pyhamcrest==2.1.0
    # via -r requirements.in
pysocks==1.7.1
    # via urllib3
pytest==8.3.2
    # via
    #   allure-pytest
    #   pytest-xdist
pytest-xdist==3.6.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via faker
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.7.24
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
rpds-py==0.20.0
    # via
    #   jsonschema
    #   referencing
selenium==4.24.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via trio
sortedcontainers==2.4.0
    # via
    #   hypothesis
    #   trio
trio==0.26.2
    # via
    #   selenium
    #   trio-websocket
trio-websocket==0.11.1
    # via selenium
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   selenium
urllib3[socks]==2.2.2
    # via
    #   requests
    #   selenium
websocket-client==1.8.0
    # via selenium
win32-setctime==1.1.0
    # via loguru
wsproto==1.2.0
    # via trio-websocket

# The following packages are considered to be unsafe in a requirements file:
# pip
